+++ eth-providers/src/utils/index.ts
@@ -5,4 +5,5 @@ export * from './transactionReceiptHelper';
 export * from './handleTxResponse';
 export * from './sendTx';
 export * from './transactionHelper';
+export * from './logs';
 export * from './utils';
